it's an int, bozo! :)
svn path=/trunk/kdebase/kdm/; revision=248287

+++ kdm/backend/genauth.c
@@ -293,7 +293,7 @@ void
 AddPreGetEntropy (void)
 {
     static long offset;
-    long readlen;
+    int readlen;
 
     AddTimerEntropy();
     if ((readlen = sumFile (randomFile, BSIZ, SEEK_SET, offset)) == BSIZ) {
